to_geoviews and to_hvplot in _scene_converters.py
As requested by David in #2106
